convert Point2D and Vector2D values to doubles, add drunk, dizzy, and sane enemy movement

+++ GalacticFX/Game/Coordinate2D.java
@@ -4,9 +4,9 @@ package Game;
  * Created by Wynn on 3/26/2017.
  */
 public class Coordinate2D {
-    private int x, y;
+    private double x, y;
 
-    public Coordinate2D(int x, int y) {
+    public Coordinate2D(double x, double y) {
         this.x = x;
         this.y = y;
     }
@@ -16,27 +16,27 @@ public class Coordinate2D {
         this.y = c.getY();
     }
 
-    public int getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public int getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
-    public void translateX(int offset) {
+    public void translateX(double offset) {
         setX(getX()+offset);
     }
 
-    public void translateY(int offset) {
+    public void translateY(double offset) {
         setY(getY()+offset);
     }
